requirements.txt
Flask==3.0.3
Flask_JWT_Extended==4.6.0
PyMySQL==1.1.1
python-dotenv==1.0.1
Werkzeug==3.0.4
